Use float64 (double) instead of float32
Replace numpy float32 with float64 across filters, rewards, connectors, and utils to standardize on double-precision arrays and type annotations. Updated NDArray type hints, np.array/concatenate dtypes, Gym Box dtype, and related function signatures in the following files: src/eprllib/Agents/Filters/BaseFilter.py, src/eprllib/Agents/Filters/DefaultFilter.py, src/eprllib/Agents/Filters/FullySharedParametersFilter.py, src/eprllib/Agents/Rewards/BaseReward.py, src/eprllib/AgentsConnectors/CentralizedConnector.py, src/eprllib/AgentsConnectors/FullySharedParametersConnector.py, src/eprllib/AgentsConnectors/HierarchicalConnector.py, src/eprllib/Utils/env_config_utils.py, src/eprllib/Utils/episode_fn_utils.py. This ensures consistent numeric precision across observations, actions, and internal processing.

+++ src/eprllib/Agents/Filters/BaseFilter.py
@@ -11,7 +11,7 @@ must be based in this class.
 """
 from eprllib import logger
 from typing import Any, Dict
-from numpy import float32
+from numpy import float64
 from numpy.typing import NDArray
 
 class BaseFilter:
@@ -35,7 +35,7 @@ class BaseFilter:
         self,
         env_config: Dict[str, Any],
         agent_states: Dict[str, Any],
-    ) -> NDArray[float32]:
+    ) -> NDArray[float64]:
         """
         Returns the filtered observations for the agent based on the environment configuration
         and agent states. This method processes the raw observations according to the filter
@@ -48,7 +48,7 @@ class BaseFilter:
             agent_states (Dict[str, Any], optional): Dictionary containing the states of the agent.
 
         Returns:
-            NDarray: Filtered observations as a numpy array of float32 values.
+            NDarray: Filtered observations as a numpy array of float64 values.
         """
         msg = "BaseFilter: This method should be implemented in a subclass."
         logger.error(msg)
