+++ src/eprllib/AgentsConnectors/HierarchicalConnector.py
@@ -10,7 +10,7 @@ The low-level agents use a fully-shared-parameter policy.
 
 import numpy as np
 from numpy.typing import NDArray
-from numpy import float32
+from numpy import float64
 from gymnasium.spaces import Box
 from gymnasium import Space
 from typing import Dict, Any, List, Tuple, Optional
@@ -233,9 +233,9 @@ class HierarchicalTwoLevelsConnector(BaseConnector):
                     dict_agents_obs[agent] = np.concatenate(
                         (
                             dict_agents_obs[agent],
-                            np.array([goals[self.top_level_agent][ix]], dtype='float32')
+                            np.array([goals[self.top_level_agent][ix]], dtype='float64')
                         ),
-                        dtype='float32'
+                        dtype='float64'
                     )
                     infos[agent].update({'goal': goals[self.top_level_agent][ix]})
                     ix += 1
@@ -245,9 +245,9 @@ class HierarchicalTwoLevelsConnector(BaseConnector):
                 dict_agents_obs[agent] = np.concatenate(
                     (
                         dict_agents_obs[agent],
-                        np.array([goals[self.top_level_agent]], dtype='float32')
+                        np.array([goals[self.top_level_agent]], dtype='float64')
                     ),
-                    dtype='float32'
+                    dtype='float64'
                 )
                 infos[agent].update({'goal': goals[self.top_level_agent]})
         
@@ -522,9 +522,9 @@ class HierarchicalThreeLevelsConnector(BaseConnector):
                     dict_agents_obs[agent] = np.concatenate(
                         (
                             dict_agents_obs[agent],
-                            np.array([goals[self.top_level_agent][ix]], dtype='float32')
+                            np.array([goals[self.top_level_agent][ix]], dtype='float64')
                         ),
-                        dtype='float32'
+                        dtype='float64'
                     )
                     infos[agent].update({'goal': goals[self.top_level_agent][ix]})
                     ix += 1
@@ -534,9 +534,9 @@ class HierarchicalThreeLevelsConnector(BaseConnector):
                 dict_agents_obs[agent] = np.concatenate(
                     (
                         dict_agents_obs[agent],
-                        np.array([goals[self.top_level_agent]], dtype='float32')
+                        np.array([goals[self.top_level_agent]], dtype='float64')
                     ),
-                    dtype='float32'
+                    dtype='float64'
                 )
                 infos[agent].update({'goal': goals[self.top_level_agent]})
         
@@ -553,7 +553,7 @@ class HierarchicalThreeLevelsConnector(BaseConnector):
         dict_agents_obs: Dict[str,Any],
         infos: Dict[str, Dict[str, Any]],
         objectives: Dict[str, List[int | float]]
-        ) -> Tuple[Dict[str,NDArray[float32]], Dict[str, Dict[str, Any]]]:
+        ) -> Tuple[Dict[str,NDArray[float64]], Dict[str, Dict[str, Any]]]:
         """
         Processes a dictionary of agent objectives to generate a combined vector.
         
@@ -607,7 +607,7 @@ class HierarchicalThreeLevelsConnector(BaseConnector):
                     dict_agents_obs[agent],
                     final_vector
                 ),
-                dtype='float32'
+                dtype='float64'
             )
             infos[agent].update({'objective': final_vector})
